Add trivial test CLI
The CLI just reads stdin line by line. If the line is valid hex, then it will be printed back on stdout in normalized (i.e. lower-case) form. If not, the program logs a descriptive message on stderr and exists with a non-zero status code.

The purpose of the tool is just to exercise the libary for test purposes.